+++ src/Customer/Exception/PasswordMismatch.php
@@ -23,9 +23,4 @@ final class PasswordMismatch extends Error
     {
         return new self('Old password does not match our records');
     }
-
-    public static function byLength(): self
-    {
-        return new self('Password does not match length requirements');
-    }
 }