+++ src/Customer/Service/CustomerRegisterInput.php
@@ -14,14 +14,14 @@ use OxidEsales\GraphQL\Base\Infrastructure\Legacy;
 use OxidEsales\GraphQL\Storefront\Customer\DataType\Customer;
 use OxidEsales\GraphQL\Storefront\Customer\Exception\CustomerExists;
 use OxidEsales\GraphQL\Storefront\Customer\Exception\InvalidEmail;
-use OxidEsales\GraphQL\Storefront\Customer\Exception\PasswordMismatch;
+use OxidEsales\GraphQL\Storefront\Customer\Exception\PasswordValidationException;
 use OxidEsales\GraphQL\Storefront\Customer\Infrastructure\CustomerRegisterFactory;
-use OxidEsales\GraphQL\Storefront\Customer\Infrastructure\Repository;
+use OxidEsales\GraphQL\Storefront\Customer\Infrastructure\RepositoryInterface;
 use TheCodingMachine\GraphQLite\Annotations\Factory;
 
 final class CustomerRegisterInput
 {
-    /** @var Repository */
+    /** @var RepositoryInterface */
     private $repository;
 
     /** @var Legacy */
@@ -31,7 +31,7 @@ final class CustomerRegisterInput
     private $customerRegisterFactory;
 
     public function __construct(
-        Repository $repository,
+        RepositoryInterface $repository,
         Legacy $legacyService,
         CustomerRegisterFactory $customerRegisterFactory
     ) {
@@ -57,7 +57,7 @@ final class CustomerRegisterInput
             strlen($password) == 0 ||
             (strlen($password) < $this->legacyService->getConfigParam('iPasswordLength'))
         ) {
-            throw PasswordMismatch::byLength();
+            throw new PasswordValidationException('Password does not match length requirements');
         }
 
         if ($this->repository->checkEmailExists($email)) {