+++ tests/Codeception/Config/params.php
@@ -36,6 +36,8 @@ return [
     'PHP_BIN'              => getenv('PHPBIN') ?: 'php',
     'SCREEN_SHOT_URL'      => getenv('CC_SCREEN_SHOTS_PATH') ?: '',
     'THEME_ID'             => getenv('THEME_ID') ?: 'twig',
+    'MAIL_HOST'            => getenv('MAIL_HOST') ?: 'mailpit',
+    'MAIL_WEB_PORT'        => getenv('MAIL_WEB_PORT') ?: '8025',
 ];
 
 function getTestDataDumpFilePath(): string