+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.1.0] - Unreleased
+## [3.1.0-rc.1] - 2024-06-12
 
 ### Added
 - `variantSelections` query for fetching multidimensional variants [PR-11](https://github.com/OXID-eSales/graphql-storefront-module/pull/11)
@@ -14,16 +14,47 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - PHP 8.2 support
 - Classes
     - `OxidEsales\GraphQL\Storefront\Manufacturer\DataType\ManufacturerImage`
+    - `OxidEsales\GraphQL\Storefront\Shared\Infrastructure\OxNewFactory`
+    - `OxidEsales\GraphQL\Storefront\Shared\Infrastructure\OxNewFactoryInterface`
+- Password forgotten feature
+  - Mutations:
+    - `customerPasswordForgotRequest`
+    - `customerPasswordReset`
+  - Classes:
+    - `OxidEsales\GraphQL\Storefront\Customer\Infrastructure\Password`
+    - `OxidEsales\GraphQL\Storefront\Customer\Infrastructure\PasswordInterface`
+    - `OxidEsales\GraphQL\Storefront\Customer\Exception\CustomerNotFoundByUpdateHash`
+    - `OxidEsales\GraphQL\Storefront\Customer\Exception\PasswordValidationException`
+    - `OxidEsales\GraphQL\Storefront\Customer\Infrastructure\RepositoryInterface`
+    - `OxidEsales\GraphQL\Storefront\Customer\Service\CustomerInterface`
+    - `OxidEsales\GraphQL\Storefront\Customer\Service\PasswordInterface`
+    - `OxidEsales\GraphQL\Storefront\Shared\Infrastructure\RepositoryInterface`
+  - Methods:
+    - `OxidEsales\GraphQL\Storefront\Customer\Infrastructure\Repository::saveNewPasswordForCustomer`
+    - `OxidEsales\GraphQL\Storefront\Customer\Infrastructure\Repository::getCustomerByPasswordUpdateHash`
+    - `OxidEsales\GraphQL\Storefront\Customer\Service\Password::sendPasswordForgotEmail`
+    - `OxidEsales\GraphQL\Storefront\Customer\Service\Password::resetPasswordByUpdateHash`
+
+### Deprecated
+- Aliases for Dependency Injection
+    - `OxidEsales\GraphQL\Storefront\Customer\Service\Customer`
+    - `OxidEsales\GraphQL\Storefront\Customer\Infrastructure\Repository`
 
 ### Changed
 - Replace webmozart/path-util usage with symfony/filesystem
 - New module logo
 - Updated the structure to Codeception 5
 - Modify github workflows to use new universal workflow
+- Use new interfaces instead of direct classes
 
 ### Removed
 - PHP 8.0 support
 - Migration trigger on module activation
+- `OxidEsales\GraphQL\Storefront\Shared\Service\Authorization`
+- `byLength`in `OxidEsales\GraphQL\Storefront\Customer\Exception\PasswordMismatch`
+
+### Fixed
+- Add error message to product response when variant loading is disabled [#0007421](https://bugs.oxid-esales.com/view.php?id=7421)
 
 ## [3.0.0] - 2023-06-08
 
@@ -168,6 +199,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `oxid-esales/graphql-account`
     - `oxid-esales/graphql-checkout`
 
+[3.1.0-rc.1]: https://github.com/OXID-eSales/graphql-storefront-module/compare/v3.0.0...v3.1.0-rc.1
 [3.0.0]: https://github.com/OXID-eSales/graphql-storefront-module/compare/v2.1.0...v3.0.0
 [2.2.0]: https://github.com/OXID-eSales/graphql-storefront-module/compare/v2.1.0...b-6.5.x
 [2.1.0]: https://github.com/OXID-eSales/graphql-storefront-module/compare/v2.0.1...v2.1.0